+++ coding_agent/config.py
@@ -18,6 +18,7 @@ class Config:
     memory_distill:bool=True
     resume_max_chars:int=6000
     memory_max_records:int=200; memory_ttl_days:int=0
+    show_todos:bool=True
     @classmethod
     def from_env(cls, workspace_arg=None):
         load_dotenv()
@@ -39,4 +40,5 @@ class Config:
           os.getenv('CODER_MEMORY_INJECT','1')!='0',int(os.getenv('CODER_MEMORY_TOPK','4')),int(os.getenv('CODER_MEMORY_MAX_CHARS','1500')),float(os.getenv('CODER_MEMORY_MIN_SCORE','0.5')),
           os.getenv('CODER_MEMORY_DISTILL','1')!='0',
           int(os.getenv('CODER_RESUME_MAX_CHARS','6000')),
-          int(os.getenv('CODER_MEMORY_MAX_RECORDS','200')),int(os.getenv('CODER_MEMORY_TTL_DAYS','0')))
+          int(os.getenv('CODER_MEMORY_MAX_RECORDS','200')),int(os.getenv('CODER_MEMORY_TTL_DAYS','0')),
+          os.getenv('CODER_TODOS','1')!='0')